requirements.txt
numpy==1.20.0
scipy==1.7.1
librosa==0.8.1
numba==0.54.0
tqdm==4.62.2
hydra-core==1.1.1
pyloudnorm==0.1.0
tensorboard==2.6.0
llvmlite==0.37.0
torch==1.9.0
webrtcvad==2.0.10
